Move pants.engine.exp.examples to pants_test.engine.v2.examples.
